Don't throw on errors, pass to client.
In many cases Active Campaign API returns 200,
but the body contains error. Those errors should be
handled by the client: throwing exception makes it impossible.

Example: contact_create: if contact is already created, it makes
no harm to try and create it again. The client can check [result_code]
and implement the logic to handle "the client already created",
or just ignore it.